unpaper: rewrite options parsing, and update documentation
The new option parsing uses getopt_long_only() so that it's mostly
compatible with the old options. Only two options have been removed:
--input-sequence and --output-sequence (which were used by some tests, they
are now updated).

With this change, options are parsed once per execution rather than once
per sheet, which means that the cost of passing multiple pages for
sequential processing is much lower.

Further improvement in file processing is to be expected.