Naive implementation of file input through libav.
By using libav to read files, we can avoid a long list of unsafe fscanf()
statements, and allow for reading more image formats (although not all of
those that would be useful).

The changes as they are implemented now are very naive and do not support
things like TIFF files or JPEG, but it alreayd allows 24-bit color PNG
files (without alpha channel). B/W PNG is not supported (yet) because of a
difference in pixel format.

Since now the "input type" of the file is no longer relevant, remove it
from the verbose output.