Make setPixel() handle B/W images by using the global absBlackThreshold.
Since now the parameter is computed at global level, no change of API is
needed for setPixel: decide whether the pixel is black or white depending
on the global threshold, and do so straight in the function.

This allows removing errOutput() calls, and collapses the MONOBLACK and
MONOWHITE cases down.