meson build: require POSIX.1-2008 conformance.
Because `strdup()` is only present in POSIX.1-2001 and later, and
explicitly not in ISO C11, enable the preprocessor macro to request
POSIX.1-2008 compliance.

Since this does not give us assurance that the function won't be
implicitly defined, add a request for error in case of implicit
pointer conversion to integer which _should_ always be an error.

Note that this is only supported for GCC 5.0 and later.