tests: use pytest (and pillow) for tests.
Instead of reusing the original Autotools-based tests and the
compare-image binary, rewrite the tests to use a pytest-compatible
suite, that can even be ran in parallel (but is not exercised as such by
meson).

The new tests are mostly identical in execution from the previous ones
with the exception of not using different output names based on the test
(because pytest gives them independent temporary directories), and also
allow referencing the same golden files, so that we need fewer of them.

Note that this requires a newer version of Meson to be able to disable
the timeout on tests, which we want since the tests themselves take
a significant amount of time, particularly when executed sequentially.