Updated image I/O functions for FFmpeg 5.0
The functions `loadImage` and `saveImage` in
`file.c` rely upon a number of deprecated functions
from FFmpeg, including some that are removed in
FFmpeg. These have been updated to the current
best practices recommended by FFmpeg.

Notably, this moves away from the use of
`avcodec_decode_video2` and `avcodec_encode_video2`
in favor of the decoupled encoding and
decoding APIs.