Create new primitives to shift points and rectangles.
A lot of the code up to now have to move x/y coordinates over
a certain pattern. Providing some primitive functions for abstractions
make it possible to pay more attention to what is done, rather
than which variable is used.

Also include a boundary check for a rectangle-in-rectangle which
is another common operation.

This also fixes the blackfilter scan to check for the *correct* vertex.