Move the remaining image handling functions to blit.c
This gets rid of tools.c altogether, as the create/replace/free image APIs
are now part of blit.

This may not be the final state for them, as they could use a separate
unit instead. But it allows reducing the legacy APIs as we move along,
and stop depending on global state variables.