Wrap around the creation of compatible temporary images.
Instead of having to get the details of AVFrame every time, allow
the creation of a new image based on the format and parameters
of another.

This allows making AVFrame opaque to *most* of the code, except
for (obviously) the image manipulation routines.

The unfortunate side effect is needing to fix all the includes that
were previously imported via libavutil.