Rework physical dimensions parsing.
This is a significant change in the parsing of physical dimension and (paper) sizes compared to previously.

PPI (formerly referred to as DPI) calculation is applied only after all options have been parsed, which means it is no longer possible to mix parameters with different PPI interpretation.

This would allow using dimension parsing more widely as it is now much more reliable, if not very optimised.

Internal calculation has moved to integer mils (1/1000th of an inch) rather than floating inches, which caused some of the golden files to ever so slightly increase in size (1px on the vertical.)

As mils is now used internally, accept mils as a valid unit (together with mm, cm, and in.)

And instead of only accepting a3/4/5 and letter/legal, include the whole A0~A10 series, but keep the more common formats higher in the table.