Rewrite edge parsign and handling.
The new Edge structure has the same size as the previous bitmask, but
simplifies the flow of the code by just accessing forur booleans instead
of using bitwise math.

The parsing is changed and made a lot stricter, allowing only the tokens
themselves instead of a horrible mix of "leftrighttopbottom" which it did
before (despite the man page saying this should be comma-separated.)

On the other hand, the parsing is now case insensitive.