Hide some of the pre-validation variables away from the processing code.
This moves the temporary variables that are used to store command line
arguments before their validations inside a block that terminates before
processing start.

It ensures that for parameters that need to be calculated or validated,
the processing code only gets access to the latter.

Note that this currently cannot be done for variables that are referenced
as pointers, but that is only the blackfilterExclude array.